wip: Use "inverted-Z" depth buffer
This version doesn't give us many benefit yet, we need
GL_EXT_clip_control (or similar) to get the extra precision.
In this PR we attempt to find and fix of the problems coming from
a reversed depth buffer.

Specifically:

- invert the depth buffer with glDepthRange() or equivalent
- clear the depth buffer with 0
- invert all depth function comparisons 
- fix all screen-space effects
  - ssao
  - DoF
- fix shadows and shadow biases


TODO:
As explained in ARB_clip_control spec, using glDepthRange to
reverse the depth buffer doesn't improve the precision.
This transformation needs to be done in the projection matrix.